requirements.txt
Flask~=3.1.1
celery~=5.5.3
django-scheduler~=0.10.1